pyproject.toml
[project]
name = "etl-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "black>=25.11.0",
    "duckdb>=1.1.3",
    "duckdb-engine>=0.13.2",
    "pandas>=2.3.3",
    "pandas-stubs>=2.3.2.250926",
    "psycopg>=3.2.12",
    "pymysql>=1.1.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "sqlalchemy-serializer>=1.4.12",
]
